+++ packages/react-virtual/e2e/app/vite.config.ts
@@ -16,6 +16,14 @@ export default defineConfig({
         ),
         'smooth-scroll': path.resolve(__dirname, 'smooth-scroll/index.html'),
         'stale-index': path.resolve(__dirname, 'stale-index/index.html'),
+        'direct-dom-updates': path.resolve(
+          __dirname,
+          'direct-dom-updates/index.html',
+        ),
+        'cached-measurements': path.resolve(
+          __dirname,
+          'cached-measurements/index.html',
+        ),
       },
     },
   },