+++ benchmarks/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tanstack/virtual-benchmarks
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-virtual@3.14.2
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-virtual@3.14.1
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`73e115d`](https://github.com/TanStack/virtual/commit/73e115d53faf01f073bc4a7bd71ee0139307f4a8)]:
+  - @tanstack/react-virtual@3.14.0
+
 ## 0.0.2
 
 ### Patch Changes